Refactor: File Flatten 리팩토링, 패키지 플래튼하게 변경
