+++ src/main/java/kr/hhplus/be/server/application/PaymentCriteria.java
@@ -1,4 +1,4 @@
-package kr.hhplus.be.server.application.dto;
+package kr.hhplus.be.server.application;
 
 import lombok.Builder;
