+++ src/main/java/kr/hhplus/be/server/application/PointCriteria.java
@@ -1,4 +1,4 @@
-package kr.hhplus.be.server.application.dto;
+package kr.hhplus.be.server.application;
 
 import kr.hhplus.be.server.domain.point.PointCommand;
 import kr.hhplus.be.server.presentation.point.object.PointRequest;
