+++ src/main/java/kr/hhplus/be/server/application/ReservationResult.java
@@ -1,4 +1,4 @@
-package kr.hhplus.be.server.application.dto;
+package kr.hhplus.be.server.application;
 
 import lombok.Builder;
