+++ src/main/java/kr/hhplus/be/server/presentation/payment/object/PaymentRequest.java
@@ -1,7 +1,7 @@
 package kr.hhplus.be.server.presentation.payment.object;
 
 import jakarta.validation.constraints.NotNull;
-import kr.hhplus.be.server.application.dto.PaymentCriteria;
+import kr.hhplus.be.server.application.PaymentCriteria;
 
 public record PaymentRequest(
         @NotNull Long reservationId
