+++ src/test/java/kr/hhplus/be/server/application/PaymentFacadeIntegrationTest.java
@@ -1,6 +1,5 @@
 package kr.hhplus.be.server.application;
 
-import kr.hhplus.be.server.application.dto.PaymentCriteria;
 import kr.hhplus.be.server.domain.payment.Payment;
 import kr.hhplus.be.server.domain.point.Point;
 import kr.hhplus.be.server.domain.reservation.Reservation;
