+++ src/test/java/kr/hhplus/be/server/application/ReservationFacadeConcurrencyTest.java
@@ -1,6 +1,5 @@
 package kr.hhplus.be.server.application;
 
-import kr.hhplus.be.server.application.dto.ReservationCriteria;
 import kr.hhplus.be.server.domain.reservation.Reservation;
 import kr.hhplus.be.server.infrastructure.repository.ReservationRepository;
 import kr.hhplus.be.server.infrastructure.repository.SeatRepository;
