+++ src/test/java/kr/hhplus/be/server/presentation/point/PointControllerTest.java
@@ -1,9 +1,9 @@
 package kr.hhplus.be.server.presentation.point;
 
 import kr.hhplus.be.server.application.PointFacade;
-import kr.hhplus.be.server.application.dto.PointResult;
-import kr.hhplus.be.server.application.dto.PointCriteria;
-import kr.hhplus.be.server.application.dto.PointHistoryResult;
+import kr.hhplus.be.server.application.PointResult;
+import kr.hhplus.be.server.application.PointCriteria;
+import kr.hhplus.be.server.application.PointHistoryResult;
 import kr.hhplus.be.server.domain.pointhistory.PointHistoryType;
 import kr.hhplus.be.server.presentation.point.object.PointHistoryResponse;
 import kr.hhplus.be.server.presentation.point.object.PointRequest;
